нельзя понижать статус самого себя, для этого можно создать другой аккаунт, поднять его до своего уровня и понизить старый аккаунт а вообще этот вопрос надо пересмотреть, по логике нельзя понижать всё, что твоего уровня и выше, управление правами серьёзная штука

+++ module/class.user.php
@@ -155,6 +155,8 @@ class user extends Module
         }
         $for_user_id = $this->data->get("for_user_id");
         if (!$this->exists($for_user_id)) return;
+        $user_id = $this->data->load("user")["id"];
+        if ($user_id == $for_user_id) return;
         $status = $this->data->get("status");
         if ($status == "drop")
             $this->delete($for_user_id);
